Replace custom/modified libircclient with libircclient-1.6
This change will definitely cause xdcc to break, but I will slow fix it with changes to libircclient. The reason to "separate" libircclient, or rather to use a release version rather than the custom version, is to enable me to

1. upgrade to the later versions easily.
2. to one day be able to use libircclient as a shared library.

Those two goals are not possible as long as xdccget relies on a very custom, very modified libircclient. However, those customizations are necessary, so I will carefully and methodically reintroduce them in such a way that it allows libircclient to remain distinct and easily removable.