Remove libircclient patches
These patches were here so that I could keep track of my changes, should a newer version of libircclient ever be released and I want to incorporate those new fixes/features, but with all of the changes that I'll need to make to libircclient (for performance and security), it seems unlikely that I will be able to easily backport a newer release.

So, let's merge libircclient properly into the repo.