Fix CodeQL alert 105
The alert is asserting that the array bounds check `offset < length - 1` does not precede the array index, `buf[offset]`. But this is, in fact, a false positive, because `offset` is bounded by the `for` condition statement, `offset < length`. Nevertheless, it's simple to fix and I think it makes the code a little more readable.