Fix IKOS analysis
In order to fix the analysis, I had to remove the reference to "config.h" as it did not exist and caused ikos to abort. Then, I supplied the right commands to ikos to use libbsd. IKOS does compile xdccget, but it does not use meson, because when I used ikos-scan ninja -C build, IKOS found nothing.