Initial commit with "new" build system.
Meson is nice, but it is an additional dependency to download/install for every platform. I was never happy about this, but it seemed like an acceptable sacrafice rather than support autotools. However, Meson also seemed overkill for my needs, which are to check for the existence of libbsd, etc.

oconfigure is simple and portable...simpler than Meson and autobuild. No extra dependencies.