Changed xdccget to use the nick format "xdccget!$PID".
This is the struggle of finding a balance between anonymity and short code, but I think this finds a good balance. Process IDs are not super random, but there are probably not many users of this xdccget. And if it ever becomes a problem, I'll add another flag to be able to override the nick.